Moving the proto + services to suuchi-examples instead of suuchi-core
