+++ suuchi-examples/src/main/scala/in/ashwanthkumar/suuchi/rpc/PingService.scala
@@ -1,7 +1,7 @@
 package in.ashwanthkumar.suuchi.rpc
 
-import in.ashwanthkumar.suuchi.rpc.generated.PingServiceGrpc
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{PingRequest, PingResponse}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.PingServiceGrpc
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{PingRequest, PingResponse}
 import io.grpc.stub.StreamObserver
 
 class PingService extends PingServiceGrpc.PingServiceImplBase {
