+++ suuchi-examples/src/main/scala/in/ashwanthkumar/suuchi/rpc/SuuchiPutService.scala
@@ -1,12 +1,11 @@
 package in.ashwanthkumar.suuchi.rpc
 
-import com.google.protobuf.ByteString
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiPutGrpc
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{PutResponse, PutRequest, GetRequest, GetResponse}
-import in.ashwanthkumar.suuchi.store.{WriteStore, ReadStore}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.PutGrpc
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{PutRequest, PutResponse}
+import in.ashwanthkumar.suuchi.store.WriteStore
 import io.grpc.stub.StreamObserver
 
-class SuuchiPutService(store: WriteStore) extends SuuchiPutGrpc.SuuchiPutImplBase {
+class SuuchiPutService(store: WriteStore) extends PutGrpc.PutImplBase {
   override def put(request: PutRequest, responseObserver: StreamObserver[PutResponse]): Unit = {
     val key = request.getKey.toByteArray
     val value = request.getValue.toByteArray
