+++ suuchi-examples/src/main/scala/in/ashwanthkumar/suuchi/rpc/SuuchiReadService.scala
@@ -1,13 +1,12 @@
 package in.ashwanthkumar.suuchi.rpc
 
 import com.google.protobuf.ByteString
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{GetRequest, GetResponse}
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiReadGrpc
+import in.ashwanthkumar.suuchi.examples.rpc.generated.{ReadGrpc, SuuchiRPC}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{GetRequest, GetResponse}
 import in.ashwanthkumar.suuchi.store.ReadStore
 import io.grpc.stub.StreamObserver
 
-class SuuchiReadService(store: ReadStore) extends SuuchiReadGrpc.SuuchiReadImplBase {
+class SuuchiReadService(store: ReadStore) extends ReadGrpc.ReadImplBase {
   override def get(request: GetRequest, responseObserver: StreamObserver[GetResponse]): Unit = {
     val key = request.getKey.toByteArray
     store.get(key) match {
