+++ suuchi-examples/src/main/scala/in/ashwanthkumar/suuchi/rpc/SuuchiScanService.scala
@@ -1,14 +1,14 @@
 package in.ashwanthkumar.suuchi.rpc
 
 import com.google.protobuf.ByteString
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{ScanRequest, ScanResponse}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.{ScanGrpc, SuuchiRPC}
 import in.ashwanthkumar.suuchi.partitioner.SuuchiHash
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{ScanRequest, ScanResponse}
-import in.ashwanthkumar.suuchi.rpc.generated.{SuuchiRPC, SuuchiScanGrpc}
 import in.ashwanthkumar.suuchi.store.{KV, Store}
 import in.ashwanthkumar.suuchi.utils.{ByteArrayUtils, ConnectionUtils}
 import io.grpc.stub.{ServerCallStreamObserver, StreamObserver}
 
-class SuuchiScanService(store: Store) extends SuuchiScanGrpc.SuuchiScanImplBase with ConnectionUtils {
+class SuuchiScanService(store: Store) extends ScanGrpc.ScanImplBase with ConnectionUtils {
 
   private def buildResponse(response: KV): ScanResponse = {
     SuuchiRPC.ScanResponse.newBuilder()
