+++ suuchi-examples/src/test/scala/in/ashwanthkumar/suuchi/rpc/PingServiceTest.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.rpc
 
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{PingRequest, PingResponse}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{PingRequest, PingResponse}
 import io.grpc.stub.StreamObserver
 import org.mockito.ArgumentCaptor
 import org.mockito.Mockito._
