+++ suuchi-examples/src/test/scala/in/ashwanthkumar/suuchi/rpc/SuuchiPutServiceTest.scala
@@ -1,7 +1,7 @@
 package in.ashwanthkumar.suuchi.rpc
 
 import com.google.protobuf.ByteString
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{PutRequest, PutResponse}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{PutRequest, PutResponse}
 import in.ashwanthkumar.suuchi.store.InMemoryStore
 import io.grpc.stub.StreamObserver
 import org.mockito.ArgumentCaptor
