+++ suuchi-examples/src/test/scala/in/ashwanthkumar/suuchi/rpc/SuuchiScanServiceTest.scala
@@ -1,7 +1,7 @@
 package in.ashwanthkumar.suuchi.rpc
 
-import in.ashwanthkumar.suuchi.rpc.generated.SuuchiRPC.{ScanRequest, ScanResponse}
-import in.ashwanthkumar.suuchi.store.{InMemoryStore, Store}
+import in.ashwanthkumar.suuchi.examples.rpc.generated.SuuchiRPC.{ScanRequest, ScanResponse}
+import in.ashwanthkumar.suuchi.store.InMemoryStore
 import io.grpc.stub.ServerCallStreamObserver
 import org.mockito.ArgumentCaptor
 import org.mockito.Mockito._
